refactor project directory structure
