+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import { ThemeProvider } from "@material-ui/styles";
 import App from "./App";
-import { generateTheme, useDarkMode } from "./theme";
+import { generateTheme, useDarkMode } from "./hook/theme";
 
 ReactDOM.render(<Index />, document.getElementById("root"));
