+++ src/typing/react-app-env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="react-scripts" />
