+++ src/util/registry_utils.test.ts
